+++ scripts/run.sh
@@ -28,15 +28,31 @@ docker run -d \
   -e CLIENTTYPE_HEADER_KEY=client-type \
   -e BACKOFFICE_HEADER_KEY=isbackoffice \
   -e USER_PROPERTIES_HEADER_KEY=userproperties \
-  -e MICRO_LC_BASE_PATH=/public/ \
-  -e PUBLIC_DIRECTORY_PATH=/usr/src/app/public \
-  -e RESOURCES_DIRECTORY_PATH=/usr/src/app/configurations \
-  -v `pwd`/.env/public:/usr/src/app/public \
-  -v `pwd`/.env/configurations:/usr/src/app/configurations \
-  -v `pwd`/.env/config.json:/usr/src/app/config.json \
+  -v `pwd`/.env/public:/usr/static/public \
+  -v `pwd`/.env/configurations:/usr/static/configurations \
+  -v `pwd`/.env/config.json:/usr/static/config.json \
   -p 3000:3000 \
   --name middleware \
-  microlc/middleware
+  microlc/middleware:3.0.0-rc1
+
+# docker run -d \
+#   -e LOG_LEVEL=debug \
+#   -e HTTP_PORT=3000 \
+#   -e MICROSERVICE_GATEWAY_SERVICE_NAME=microservice-gateway \
+#   -e USERID_HEADER_KEY=userid \
+#   -e GROUPS_HEADER_KEY=usergroups \
+#   -e CLIENTTYPE_HEADER_KEY=client-type \
+#   -e BACKOFFICE_HEADER_KEY=isbackoffice \
+#   -e USER_PROPERTIES_HEADER_KEY=userproperties \
+#   -e MICRO_LC_BASE_PATH=/public/ \
+#   -e PUBLIC_DIRECTORY_PATH=/usr/src/app/public \
+#   -e RESOURCES_DIRECTORY_PATH=/usr/src/app/configurations \
+#   -v `pwd`/.env/public:/usr/src/app/public \
+#   -v `pwd`/.env/configurations:/usr/src/app/configurations \
+#   -v `pwd`/.env/config.json:/usr/src/app/config.json \
+#   -p 3000:3000 \
+#   --name middleware \
+#   microlc/middleware
 
 if command -v xclip > /dev/null
 then